Moved logic in python module for slack export
