+++ .gitignore
@@ -1,2 +1,3 @@
 # Run scripts are likely to contain API tokens
 resources/jira/run*.sh
+__pycache__
