+++ resources/jira/jnra/LICENSE
@@ -1,0 +1,2 @@
+This package is unlicensed.
+This makes it a private package.
