+++ resources/jira/jnra/setup.cfg
@@ -1,0 +1,24 @@
+[metadata]
+name = jira-nodered-automation
+version = 0.0.1
+author = C. Schalkwijk
+author_email = coen@navelpluis.nl
+description = Jira NodeRed automation tools
+long_description = file: README.md
+long_description_content_type = text/markdown
+#url = https://github.com/pypa/sampleproject
+#project_urls =
+    #Bug Tracker = https://github.com/pypa/sampleproject/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved ::Unlicensed
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
