+++ resources/jira/jnra/src/jira_nodered_automation.egg-info/PKG-INFO
@@ -1,0 +1,18 @@
+Metadata-Version: 2.1
+Name: jira-nodered-automation
+Version: 0.0.1
+Summary: Jira NodeRed automation tools
+Home-page: UNKNOWN
+Author: C. Schalkwijk
+Author-email: coen@navelpluis.nl
+License: UNKNOWN
+Platform: UNKNOWN
+Classifier: Programming Language :: Python :: 3
+Classifier: License :: OSI Approved ::Unlicensed
+Classifier: Operating System :: OS Independent
+Requires-Python: >=3.6
+Description-Content-Type: text/markdown
+License-File: LICENSE
+
+UNKNOWN
+
