+++ resources/jira/jnra/src/jira_nodered_automation.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+LICENSE
+pyproject.toml
+setup.cfg
+src/jira_nodered_automation.egg-info/PKG-INFO
+src/jira_nodered_automation.egg-info/SOURCES.txt
+src/jira_nodered_automation.egg-info/dependency_links.txt
+src/jira_nodered_automation.egg-info/top_level.txt
+src/jnra/__init__.py
+src/jnra/jiraSDK.py
+src/jnra/jira_report.py
+src/jnra/progress_monitor.py
+src/jnra/slackSDK.py
