+++ resources/jira/jnra/src/jira_nodered_automation.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+jnra
