+++ resources/jira/jnra/src/jnra/slackSDK.py
